refactor(GraphQL): Move /server/graphql to /imports plugin
