+++ imports/plugins/core/graphql/server/resolvers/ping.js
