--- private/plugins/README.md
@@ -1,4 +1,0 @@
-### Do not create files in this directory manually!
-
-The content of this directory is generated automatically by the
-Reaction asset builder and will be overwritten at each startup.
